feat: optimize tailwind theme eject

+++ bin/generate-tailwind-theme.js
@@ -6,18 +6,30 @@ const prettier = require('prettier');
 const resolveConfig = require('tailwindcss/resolveConfig');
 const tailwindConfig = require('../tailwind.config');
 
-const { theme } = resolveConfig(tailwindConfig);
-const themeStr = JSON.stringify(theme.colors);
-const ts = `
-  export const colors = ${themeStr};
-`;
+const themePath = path.resolve(__dirname, '../src/theme.ts');
+const tailwindConfigPath = path.resolve(__dirname, '../tailwind.config.js');
 
-try {
-  fs.writeFileSync(
-    path.resolve(__dirname, '../src/theme.ts'),
-    prettier.format(ts, { parser: 'babel' }),
-    'utf-8',
-  );
-} catch (err) {
-  console.log(err.message);
+if (
+  !fs.existsSync(themePath) ||
+  fs.statSync(themePath).mtime < fs.statSync(tailwindConfigPath).mtime
+) {
+  console.log('ejecting tialwind theme...');
+
+  const { theme } = resolveConfig(tailwindConfig);
+  const themeStr = JSON.stringify(theme.colors);
+  const ts = `
+    export const colors = ${themeStr};
+  `;
+
+  try {
+    fs.writeFileSync(
+      path.resolve(__dirname, '../src/theme.ts'),
+      prettier.format(ts, { parser: 'babel' }),
+      'utf-8',
+    );
+  } catch (err) {
+    console.log(err.message);
+  }
+
+  console.log('ejecting tialwind theme complete.');
 }
